+++ CHANGELOG.md
@@ -386,3 +386,4 @@
 - 2024-05-01: Bench qualify, descend, follow up, user, +2
 - 2024-05-03: Impl mailbox, follow up, logo, mod, +2 query
 - 2024-05-03: Docs user, 0001 initial,  outreach base, bettercontact, +2
+- 2024-05-04: Valid mutate, follow up, user, descend, +4